+++ collective/mediaelementjs/events.py
@@ -1,11 +1,11 @@
 from zope.cachedescriptors import property
 from zope.interface import alsoProvides, noLongerProvides
 
-from collective.mediaelementjs.interfaces import IVideo, IMediaInfo
+from collective.mediaelementjs.interfaces import IVideo, IMediaInfo, IAudio
 from collective.mediaelementjs.metadata_extraction import parse_raw
 from collective.mediaelementjs.metadata_extraction import defensive_get
 
-from Products.ATContentTypes import interface
+from Products.ATContentTypes.interfaces.file import IFileContent
 from Products.Archetypes.interfaces import IObjectInitializedEvent
 
 from StringIO import StringIO
@@ -19,12 +19,22 @@ VIDEO_EXTENSIONS = [
     '.wmv',
     '.mpeg', '.mpg', '.mpe',
     '.mov', '.qt',
-    ]
+]
+
+AUDIO_EXTENSIONS = [
+    '.mp3',
+    '.ogg', '.oga',
+    '.m4a',
+    '.wav', '.wma', '.webma'
+]
 
 def remove_marker(object):
     changed = False
     if IVideo.providedBy(object):
         noLongerProvides(object, IVideo)
+        changed = True
+    if IAudio.providedBy(object):
+        noLongerProvides(object, IAudio)
         changed = True
     if changed:
         object.reindexObject(idxs=['object_provides'])
@@ -57,8 +67,14 @@ class ChangeView(object):
         if ext in VIDEO_EXTENSIONS:
             self.handleVideo()
 
+        if ext in AUDIO_EXTENSIONS:
+            self.handleAudio()
+
     def check_extension(self):
         for ext in VIDEO_EXTENSIONS:
+            if self.filename.endswith(ext):
+                return ext
+        for ext in AUDIO_EXTENSIONS:
             if self.filename.endswith(ext):
                 return ext
         return None
@@ -67,9 +83,23 @@ class ChangeView(object):
         handle = self.file_handle
         metadata = parse_raw(handle)
         handle.close()
-        
+
         if not IVideo.providedBy(self.content):
             alsoProvides(self.content, IVideo)
+            self.object.reindexObject(idxs=['object_provides'])
+
+        info = IMediaInfo(self.content)
+        info.height = defensive_get(metadata, 'height')
+        info.width = defensive_get(metadata, 'width')
+        info.duration = defensive_get(metadata, 'duration')
+
+    def handleAudio(self):
+        handle = self.file_handle
+        metadata = parse_raw(handle)
+        handle.close()
+
+        if not IAudio.providedBy(self.content):
+            alsoProvides(self.content, IAudio)
             self.object.reindexObject(idxs=['object_provides'])
 
         info = IMediaInfo(self.content)
@@ -80,7 +110,7 @@ class ChangeView(object):
 
 class ChangeFileView(ChangeView):
 
-    interface = interface.IFileContent
+    interface = IFileContent
 
     @property.Lazy
     def value(self):