+++ docs/HISTORY.rst
@@ -10,8 +10,11 @@ History
  * Ignore file for git.
  * Make sure all the needed files are packaged by distutils.
  * Remove jquery from browser resource [toutpt]
- * Update mediaelement to 2.11.2
+ * Update mediaelement to 2.11.2 [toutpt]
  * Use link to display CSS stylesheet as it is the default in  Plone4 [toutpt]
+ * Add audio support [toutpt]
+ * Move audio and video init in ++resource++collective.mediaelement.js 
+ to support for all audio/video tags in the page. [toutpt]
 
 0.1.5 - 2012-06-28
 ==================